feat(datastructure): add Set data type (#40)
* feat: add set data structure

This commit introduces a new data structure, Set, which provides the basic operations of a set data structure:
- Add an element to the set.
- Check if an element exists in the set.
- Delete an element from the set.
- Get a slice of all elements in the set.

* feat(set): support multiple values add/delete

Adds support for adding and deleting multiple values to and from a set.

This change allows for more efficient handling of multiple values, as it eliminates the need to perform individual operations for each value.

* test(datastructure): add unit tests for SetInterface

This commit adds unit tests for the SetInterface to ensure that it is working as expected.

The tests cover the following scenarios:

- Adding a single item to the set
- Adding multiple items to the set
- Checking if an item exists in the set
- Deleting an item from the set
- Getting a slice of the items in the set

* feat: clear method in Set data structure

Adds `Clear` method to `Set` data structure to remove all values from the set.

* test(set): add slice, delete and exists modes

This commit introduces three new modes to the set data structure: slice, delete, and exists.

The slice mode allows the user to pass a slice of values to the set, and the set will slice the elements.
The delete mode allows the user to pass a slice of values to the set, and the set will delete the elements.
The exists mode allows the user to pass a single value to the set, and the set will return whether or not the value exists in the set.

* feat(datastructure): add Set data type

This commit introduces a new data structure, Set, which is similar to `slice` or `array` but does not allow duplicate elements.

The Set data type can be created using the `NewSet` function and can be used to store any type of comparable element.