added function to compare element of 2 slices with ignored order of element (#41)
* feat: introduce utility to compare slice elements

Introduce a new utility `IsElementsEquals` to compare elements in two slices.

This utility is used to compare the elements of two slices regardless of their order.
It can be used to verify that two slices contain the same elements, even if they are not in the same order.

* feat(collections): check if elements of 2 slices are equal with ignored order

Adds a new function `IsElementsEquals` to the `collections` package that checks if the elements of two slices are equal, ignoring the order of the elements. This can be useful for comparing two slices that contain the same elements but in a different order.