feat(concurrency): implement concurrency package
Implement concurrency package to help execute code asynchronously.

This package provides a way to execute functions asynchronously using a worker pool.
It allows for custom configuration of the maximum number of workers, as well as the ability to use custom `sync.WaitGroup` and `sync.Mutex` instances.
It also provides a way to add errors that may occur during execution, which are returned when the `Do` method is called.