feat: added lib to implement concurrency processing (#48)
* feat(concurrency): implement concurrency package

Implement concurrency package to help execute code asynchronously.

This package provides a way to execute functions asynchronously using a worker pool.
It allows for custom configuration of the maximum number of workers, as well as the ability to use custom `sync.WaitGroup` and `sync.Mutex` instances.
It also provides a way to add errors that may occur during execution, which are returned when the `Do` method is called.

* feat: error handling in concurrency interface

Add error handling capability to the concurrency interface.

This allows you to add errors to the interface, which will be returned at Do() method if exists. It is recommended to use errors.NewWithCode() to add errors.