feat(concurrency): add func that will be run async at goroutine
This commit adds a new function to the concurrency package that will be run asynchronously at goroutine. This function already calls c.Done() after the process is complete, which simplifies the usage of the concurrency package.

BREAKING CHANGE: The signature of the AddFunc function has changed. The new signature is:

```
func AddFunc(fn func(ctx context.Context, c Interface))
```