feat: clone concurrency with empty function list
This commit introduces a `Clone` function to the `Concurrency` interface. The `Clone` function allows for the creation of a new `Concurrency` object with the same max worker as the original object, but with an empty list of functions.

This is useful in cases where you want to create a new `Concurrency` object with the same max worker as an existing object, but you do not want to copy over the list of functions from the original object.