feat(datastructure): introduce `Find` function to Set
This function allows to find the first element in the set that matches the given condition, or returns an error if no such element exists.

This is a useful addition to the set data structure, as it allows to easily find elements in the set without having to iterate over all the elements.