strformat (#66)
* docs: alias for Tmpl and TmplWithoutErr

Adds a new `T` alias for `Tmpl` and a `TE` alias for `TEmplWithoutErr`. These aliases provide a shorter, more convenient way to access these functions.

* refactor: rename Tmpl to T

This commit renames `Tmpl` to `T` to improve conciseness.

The old name `Tmpl` was long and verbose, while the new name `T` is short and easy to remember. This change also brings the naming convention in line with other functions in the package.