feat(collections): add function IsElementsEquals (#69)
Adds a new function `IsElementsEquals` to the `collections` package that checks if two slices of strings have the same elements.

This function is useful for comparing the elements of two slices without regard to their order. It can be used to check if two sets of strings are equal, or to find the intersection or union of two sets of strings.