feat: implement option for sharing locker (#103)
* feat(concurrency): Add CloneOptions to concurrency Clone method for shared locking and worker pool

This commit introduces `CloneOptions` to the `Clone` method of the `concurrency` struct, allowing for more flexible control over the cloning process. The `CloneOptions` struct contains two boolean fields:

- `LockSharing`: Determines whether the cloned concurrency object should share the same locker as the original. If true, both objects will contend for the same lock.
- `MaxWorkerSharing`: Determines whether the cloned concurrency object should share the same max worker count as the original. If true, the cloned object will have the same maximum number of workers.

The `Clone` method now accepts a `CloneOptions` struct as input. Based on the values in this struct, the cloned object can be configured to share the lock and/or max worker count with the original object. If options are not specified, the new concurrency object will have default max worker and locker.

* feat(dev): Improve development environment with new packages and extensions

This commit enhances the development environment by adding the `gh` CLI tool to the development dependencies. It also integrates two new VS Code extensions: `bernabe.gemcommit` for generating conventional commit messages and `Joundy.lumpat` for improved git workflow. These changes aim to streamline development processes and improve code quality.